+++ Categories/Oils.lua
@@ -2,5 +2,8 @@ local _, ns = ...
 ns.CATEGORIES = ns.CATEGORIES or {}
 table.insert(ns.CATEGORIES, {
     name  = "Oils",
-    items = {}
+    items = {
+        { id = 243733, rank = 1, qty = 1, enabled = true }, -- Thalassian Phoenix Oil R1
+        { id = 243734, rank = 2, qty = 1, enabled = true }, -- Thalassian Phoenix Oil R2
+    }
 })