+++ Categories/VantusRunes.lua
@@ -1,6 +1,0 @@
-local _, ns = ...
-ns.CATEGORIES = ns.CATEGORIES or {}
-table.insert(ns.CATEGORIES, {
-    name  = "Vantus Runes",
-    items = {}
-})